// src/opamp/generated/opamp_pb.d.ts
export declare type AgentRemoteConfig = Message<"opamp.proto.AgentRemoteConfig"> & {
  /**
   * Agent config offered by the management Server to the Agent instance. SHOULD NOT be
   * set if the config for this Agent has not changed since it was last requested (i.e.
   * AgentConfigRequest.last_remote_config_hash field is equal to
   * AgentConfigResponse.config_hash field).
   *
   * @generated from field: opamp.proto.AgentConfigMap config = 1;
   */
  config?: AgentConfigMap;

  /**
   * Hash of "config". The Agent SHOULD include this value in subsequent
   * RemoteConfigStatus messages in the last_remote_config_hash field. This in turn
   * allows the management Server to identify that a new config is available for the Agent.
   *
   * This field MUST be always set if the management Server supports remote configuration
   * of agents.
   *
   * Management Server must choose a hashing function that guarantees lack of hash
   * collisions in practice.
   *
   * @generated from field: bytes config_hash = 2;
   */
  configHash: Uint8Array;
};

/**
 * Describes the message opamp.proto.AgentRemoteConfig.
 * Use `create(AgentRemoteConfigSchema)` to create a new message.
 */
export declare const AgentRemoteConfigSchema: GenMessage<AgentRemoteConfig>;

/**
 * @generated from message opamp.proto.AgentConfigMap
 */
export declare type AgentConfigMap = Message<"opamp.proto.AgentConfigMap"> & {
  /**
   * Map of configs. Keys are config file names or config section names.
   * The configuration is assumed to be a collection of one or more named config files
   * or sections.
   * For agents that use a single config file or section the map SHOULD contain a single
   * entry and the key may be an empty string.
   *
   * @generated from field: map<string, opamp.proto.AgentConfigFile> config_map = 1;
   */
  configMap: { [key: string]: AgentConfigFile };
};

/**
 * Describes the message opamp.proto.AgentConfigMap.
 * Use `create(AgentConfigMapSchema)` to create a new message.
 */
export declare const AgentConfigMapSchema: GenMessage<AgentConfigMap>;

/**
 * @generated from message opamp.proto.AgentConfigFile
 */
export declare type AgentConfigFile = Message<"opamp.proto.AgentConfigFile"> & {
  /**
   * Config file or section body. The content, format and encoding depends on the Agent
   * type. The content_type field may optionally describe the MIME type of the body.
   *
   * @generated from field: bytes body = 1;
   */
  body: Uint8Array;

  /**
   * Optional MIME Content-Type that describes what's in the body field, for
   * example "text/yaml".
   *
   * @generated from field: string content_type = 2;
   */
  contentType: string;
};